+++ src/test/java/ru/yandex/practicum/sleeptracker/SleepTrackerAppTest.java
@@ -12,8 +12,8 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 
 public class SleepTrackerAppTest {
 
-    private static final int ZERO = 0;
-    private static final int     ONE = 1;
+    private static final long ZERO = 0L;
+    private static final long ONE = 1L;
 
     @Test
     void should_ReturnZeroForEmptyList() {
